// src/api/relay/generated/RepoIdQuery.graphql.ts
export type RepoIdQuery$variables = {
  repoId: string;
};
export type RepoIdQuery$data = {
  readonly node:
    | {
        readonly description?: string | null | undefined;
        readonly forkCount?: number;
        readonly licenseInfo?:
          | {
              readonly name: string;
            }
          | null
          | undefined;
        readonly nameWithOwner?: string;
        readonly owner?: {
          readonly avatarUrl: any;
          readonly login: string;
        };
        readonly primaryLanguage?:
          | {
              readonly name: string;
            }
          | null
          | undefined;
        readonly stargazerCount?: number;
      }
    | null
    | undefined;
};
export type RepoIdQuery = {
  response: RepoIdQuery$data;
  variables: RepoIdQuery$variables;
};